backend2/requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
pytz
schedule
scipy
pymongo
openai
twilio